Increase loading speed of person detail page'
